requirements.txt
Django==4.1.5
docx==0.2.4
Flask_Cors==4.0.0
gensim==4.3.1
mysql_connector_repackaged==0.3.1
numpy==1.24.3
openpyxl==3.1.2
pandas==1.1.5
py2neo
PyMySQL==1.0.2
python_docx==0.8.11
pytorch_crf==0.7.2
PyYAML==5.4.1
Requests==2.31.0
scikit_learn
scipy
seqeval==1.2.2
torch==2.0.1
tqdm==4.61.0
transformers==4.18.0
ts==0.5.1
urllib3==1.24.3
xlrd==1.2.0
XlsxWriter==3.1.1
